fix compilation issues

+++ onnx_extended/validation/cuda/cuda_fpemu.cuh
@@ -8,7 +8,7 @@ enum FpemuMode {
   E4M3_RNE = 1,
 };
 
-void fpemu_cuda_forward(const int size, const float *input, uint8_t *output,
+void fpemu_cuda_forward(const int size, const float *input, float *output,
                         FpemuMode mode, bool inplace, float scale,
                         bool block_norm, int block_size, int cuda_device);
